#35 - refactor: wrpperComponent 코드 분리
